Addressed bug with notes view not rendering multiple verses. (#10)
Context:
When viewing notes, it has been observed that a bug occures when
retrieving notes for a tag under which there are a lot of passages.

Problem:
Usually when trying to access items in an array or object which could
possibly be undefined or null in JavaScript runtime we would get an
error.

Solution:
This is where optional chaining would be an appropriate fix.

Co-authored-by: emmanuel okuchukwu <emmanuelokuchukwu@emmanuels-MacBook-Pro.local>